851 - Made fixes to how pyclamd scans files after testing on server
Needed to scan path that temporary file is written to, cannot test on 
the InMemoryUploadedFile that is f = request.FILES['attachment'].
Still need to decide what to do if we test a virus and need to test 
manual upload of EICAR test file. (The test was done on server hardcoded
but should be replicated with test upload.)
Also commented out scanning code in models, pretty sure we don't need to
scan here.